+++ lib/presentation/routes/go_router.dart
@@ -3,9 +3,9 @@ import 'package:roadmap/core/constants/routes.dart';
 import 'package:roadmap/core/utils/helper/route_helpers.dart';
 import 'package:roadmap/presentation/views/auth/login_screen.dart';
 import 'package:roadmap/presentation/views/auth/signup_screen.dart';
-import 'package:roadmap/presentation/views/home/add_screen.dart';
-import 'package:roadmap/presentation/views/home/detail_screen.dart';
-import 'package:roadmap/presentation/views/home/edit_screen.dart';
+import 'package:roadmap/presentation/views/home/goal/add_screen.dart';
+import 'package:roadmap/presentation/views/home/goal/detail_screen.dart';
+import 'package:roadmap/presentation/views/home/goal/edit_screen.dart';
 import 'package:roadmap/presentation/views/home/home_screen.dart';
 import 'package:roadmap/presentation/views/onboarding/onboarding_screen.dart';
 import 'package:roadmap/presentation/views/splash/splash_screen.dart';
