+++ lib/presentation/views/welcome/welcome_screen.dart
@@ -6,8 +6,8 @@ import 'package:roadmap/core/constants/image_strings.dart';
 import 'package:roadmap/core/constants/sizes.dart';
 import 'package:roadmap/core/constants/text_strings.dart';
 import 'package:roadmap/core/di/providers.dart';
-import 'package:roadmap/presentation/views/home_screen.dart';
-import 'package:roadmap/presentation/views/splash_screen.dart';
+import 'package:roadmap/presentation/views/home/home_screen.dart';
+import 'package:roadmap/presentation/views/splash/splash_screen.dart';
 
 class WelcomeScreen extends HookConsumerWidget {
   const WelcomeScreen({super.key});
